set validations

+++ deleteDB.sh
@@ -4,13 +4,19 @@ deleteDB()
 {
 	echo "Enter Database Name that you want to Delete"
 	read dbname
-	if [[ -d "/usr/ply/$dbname" ]]
-    	then
-		sudo rm -r /usr/ply/$dbname
-        	echo "Database Deleted Successfully" 
-        	showMenu
-    	else
-        	echo "There is no Database with name $dbname"
+	if [ -n "$dbname" ]
+       	then
+		if [[ -d "/usr/ply/$dbname" ]]
+    		then
+			sudo rm -r /usr/ply/$dbname
+        		echo "Database Deleted Successfully" 
+        		showMenu
+    		else
+        		echo "There is no Database with name $dbname"
+        		deleteDB
+    		fi
+	else
+		echo "There is no Database with name $dbname"
         	deleteDB
-    	fi
+	fi
 }
